refactored abstract class resolution to __init__.py
