+++ serivices/__init__.py
@@ -1,3 +1,9 @@
+# Abstract classes for the services (for interface reference)
+from .abstract_age_group import AbstractAgeGroupService
+from .abstract_score_service import AbstractScoreService
+from .abstract_season_service import AbstractSeasonService
+from .abstract_swimmer_service import AbstractSwimmerService
+
 from .age_group_service import AgeGroupService
 from .swimmer_service import SwimmerService
 from .season_service import SeasonService
